requirements.txt
flask
requests
PyPDF2
python-docx
flask-cors